requirements.txt
Pillow
numpy
torch
torchvision
tqdm
albumentations
tensorboard
pytorch-lightning
